M4L3 Mastery Assessment segment
This is the third Mastery for module 4 where luckily there are only two questions i had to work on.

For the first Question, I put in a module random contains functions that generate random numbers in a variety of ways. The function randint() returns an integer in the range you provide. the following code returns a number between 1 and 6: which was "from random import randint" . For this question I had to make a class Die with one attribute called sides, which has a default value of 6. Then I had to write a method called roll_die() that prints a random number between 1 and the number of sides the die has. Finally I had to made a 6-sided die and roll it 10 times.

For question 2, Similarly to Question 1I  put in a module random contains functions that generate random numbers in a variety of ways except for without the function randint() i used choice. From what i setup i allowed a possibilities of 15 choices of a combination of both numbers & letters. Using the choice() I  setup a structure that will select four items from that list in random then printing out the four chosen items.